requirements.txt
asgiref==3.11.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
